Multiplayer Tic Tac Toe game, based on server-client network and peer to peer connection.
I have designed a multiplayer Tic-Tac-Toe game using C socket
programming in which a Server creates the game in Local Area Network.
The players in LAN can connect to the server using IP address of server
provided. The connection to the server will be a client-server approach
whereas the players will be playing with each other in a Peer-to-Peer
fashion.